Fixed System Explorer Label Providers
The system explorer label providers didn't correctly show system names
and the type lib icon.

Furthermore now the type lib icon is also showing error and warning
feedback if children have errors or warnings.
