+++ plugins/org.eclipse.fordiac.ide.systemmanagement.ui/plugin.xml
@@ -391,11 +391,11 @@
             activeByDefault="true"
             contentProvider="org.eclipse.fordiac.ide.systemmanagement.ui.providers.SystemExplorerRootContentProvider"
             id="org.eclipse.fordiac.ide.systemmanagement.ui.systemexplorer.rootcontent"
-            labelProvider="org.eclipse.fordiac.ide.systemmanagement.ui.systemexplorer.SystemLabelProvider"
+            labelProvider="org.eclipse.fordiac.ide.systemmanagement.ui.systemexplorer.SystemExplorerLabelProvider"
             name="System Content"
             priority="high">
             <possibleChildren>
-                 <instanceof value="org.eclipse.core.resources.IProject"/>
+                 <instanceof value="org.eclipse.core.resources.IProject"/>                 
 	        </possibleChildren> 
        </navigatorContent>
         <navigatorContent
@@ -407,19 +407,16 @@
               priority="high">
            <possibleChildren>
               <or>
-                 <instanceof
-                       value="org.eclipse.emf.ecore.EObject">
-                 </instanceof>
-                 <instanceof
-                       value="org.eclipse.emf.edit.provider.ItemProviderAdapter">
-                 </instanceof>
+                 <!-- this possible child flag is need so that our label provider is styling the icon and text of sys files and the typeLib -->
+              	 <instanceof value="org.eclipse.core.resources.IFile" />
+              	 <instanceof value="org.eclipse.core.resources.IFolder"/> 
+                 <instanceof value="org.eclipse.emf.ecore.EObject" />
+                 <instanceof value="org.eclipse.emf.edit.provider.ItemProviderAdapter" />
               </or>
            </possibleChildren>
            <triggerPoints>
               <and>
-                 <instanceof
-                       value="org.eclipse.core.resources.IFile">
-                 </instanceof>
+                 <instanceof value="org.eclipse.core.resources.IFile" />
                  <test
                        forcePluginActivation="true"
                        property="org.eclipse.core.resources.extension"
